Reduce bit width to 16
JS was running into some sign problem with 32 bits. I couldn't mask
out the first bit.

+++ src/gol.rs
@@ -5,7 +5,7 @@ use bitvec::prelude::*;
 #[wasm_bindgen]
 pub struct Universe {
     edge_size: usize,
-    cells: BitVec<Msb0, u32>
+    cells: BitVec<Msb0, u16>
 }
 
 impl fmt::Display for Universe {
@@ -33,7 +33,7 @@ impl Universe {
         self.edge_size
     }
 
-    pub fn cells(&self) -> *const u32 {
+    pub fn cells(&self) -> *const u16 {
         self.cells.as_ptr()
     }
 
@@ -103,7 +103,7 @@ impl Universe {
 
 
     pub fn build_universe(edge_size: usize) -> Universe {
-        let cells = bitvec![Msb0, u32; 0; edge_size * edge_size];
+        let cells = bitvec![Msb0, u16; 0; edge_size * edge_size];
         Universe { edge_size, cells }
     }
 }
